+++ lib/src/call.dart
@@ -135,7 +135,7 @@ class _ZegoUIKitPrebuiltCallState extends State<ZegoUIKitPrebuiltCall>
 
     ZegoUIKit().getZegoUIKitVersion().then((version) {
       ZegoLoggerService.logInfo(
-        'version: zego_uikit_prebuilt_call:3.9.2; $version',
+        'version: zego_uikit_prebuilt_call:3.10.1; $version',
         tag: 'call',
         subTag: 'prebuilt',
       );