+++ lib/src/call_invitation/call_invitation_service.dart
@@ -175,7 +175,7 @@ class ZegoUIKitPrebuiltCallInvitationService
   }) async {
     ZegoUIKit().getZegoUIKitVersion().then((uikitVersion) {
       ZegoLoggerService.logInfo(
-        'versions: zego_uikit_prebuilt_call:3.9.2; $uikitVersion',
+        'versions: zego_uikit_prebuilt_call:3.10.1; $uikitVersion',
         tag: 'call',
         subTag: 'call invitation service',
       );