+++ lib/src/components/audio_video_view_foreground.dart
@@ -29,6 +29,7 @@ class ZegoAudioVideoForeground extends StatelessWidget {
     if (user == null) {
       return Container(color: Colors.transparent);
     }
+
     return LayoutBuilder(builder: (context, constraints) {
       return Container(
         padding: const EdgeInsets.all(5),
@@ -36,6 +37,7 @@ class ZegoAudioVideoForeground extends StatelessWidget {
           children: [
             Positioned(
               bottom: 0,
+              left: 0,
               right: 0,
               child: Container(
                 padding: const EdgeInsets.symmetric(vertical: 3, horizontal: 6),
@@ -48,7 +50,7 @@ class ZegoAudioVideoForeground extends StatelessWidget {
                   children: [
                     userName(
                       context,
-                      constraints.maxWidth * 0.8,
+                      constraints.maxWidth * 0.5,
                     ),
                     microphoneStateIcon(),
                     cameraStateIcon(),
@@ -70,6 +72,7 @@ class ZegoAudioVideoForeground extends StatelessWidget {
             ),
             child: Text(
               user?.name ?? '',
+              textAlign: TextAlign.left,
               overflow: TextOverflow.ellipsis,
               style: TextStyle(
                 fontSize: 24.0.zR,